requirements-api.txt
# Neural API Dependencies

# Core API framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Async task queue
celery>=5.3.0
redis>=5.0.0
flower>=2.0.0

# Authentication & Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6

# HTTP client for webhooks
requests>=2.31.0

# Database (optional, for production)
sqlalchemy>=2.0.0
alembic>=1.12.0
psycopg2-binary>=2.9.0  # PostgreSQL support

# Monitoring & Logging
prometheus-client>=0.18.0
python-json-logger>=2.0.0
